# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KPI Command Center",
    "description": "Operaciones ATM, sucursales y planificación geoespacial. Importes en DOP.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/competitors/institutions": {
      "get": {
        "tags": [
          "Competencia"
        ],
        "summary": "Institutions",
        "operationId": "institutions_api_v1_competitors_institutions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InstitutionOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Competencia"
        ],
        "summary": "Create Institution",
        "operationId": "create_institution_api_v1_competitors_institutions_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstitutionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstitutionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/competitors/locations": {
      "get": {
        "tags": [
          "Competencia"
        ],
        "summary": "Locations",
        "operationId": "locations_api_v1_competitors_locations_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LocationType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "provincia",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provincia"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            }
          },
          {
            "name": "hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CompetitorOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Competencia"
        ],
        "summary": "Create Location",
        "operationId": "create_location_api_v1_competitors_locations_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetitorCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/competitors/snapshots": {
      "get": {
        "tags": [
          "Competencia"
        ],
        "summary": "Snapshots",
        "operationId": "snapshots_api_v1_competitors_snapshots_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "provincia",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provincia"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            }
          },
          {
            "name": "hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MarketOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Competencia"
        ],
        "summary": "Create Snapshot",
        "operationId": "create_snapshot_api_v1_competitors_snapshots_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/competitors/comparison": {
      "get": {
        "tags": [
          "Competencia"
        ],
        "summary": "Comparison Matrix",
        "operationId": "comparison_matrix_api_v1_competitors_comparison_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provincia",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provincia"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            }
          },
          {
            "name": "hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Series_Comparison_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Comparison": {
        "properties": {
          "institucion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Institucion Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "es_propia": {
            "type": "boolean",
            "title": "Es Propia"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "fecha_corte": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Corte"
          },
          "total_atms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Atms"
          },
          "total_sucursales": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Sucursales"
          },
          "depositos": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depositos"
          },
          "prestamos": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prestamos"
          },
          "transacciones": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transacciones"
          },
          "utilizacion_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utilizacion Pct"
          },
          "cuota_depositos_muestra_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuota Depositos Muestra Pct"
          },
          "provincias_reportadas": {
            "type": "integer",
            "title": "Provincias Reportadas"
          },
          "fuentes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fuentes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "institucion_id",
          "nombre",
          "es_propia",
          "color",
          "fecha_corte",
          "total_atms",
          "total_sucursales",
          "depositos",
          "prestamos",
          "transacciones",
          "utilizacion_pct",
          "cuota_depositos_muestra_pct",
          "provincias_reportadas",
          "fuentes"
        ],
        "title": "Comparison"
      },
      "CompetitorCreate": {
        "properties": {
          "institucion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Institucion Id"
          },
          "codigo": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Codigo"
          },
          "nombre": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nombre"
          },
          "provincia": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Provincia"
          },
          "tipo": {
            "$ref": "#/components/schemas/LocationType"
          },
          "latitud": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitud"
          },
          "longitud": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitud"
          },
          "fuente": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Fuente"
          },
          "fecha_verificacion": {
            "type": "string",
            "format": "date",
            "title": "Fecha Verificacion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "institucion_id",
          "codigo",
          "nombre",
          "provincia",
          "tipo",
          "latitud",
          "longitud",
          "fuente",
          "fecha_verificacion"
        ],
        "title": "CompetitorCreate"
      },
      "CompetitorOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "institucion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Institucion Id"
          },
          "codigo": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Codigo"
          },
          "nombre": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nombre"
          },
          "provincia": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Provincia"
          },
          "tipo": {
            "$ref": "#/components/schemas/LocationType"
          },
          "latitud": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitud"
          },
          "longitud": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitud"
          },
          "fuente": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Fuente"
          },
          "fecha_verificacion": {
            "type": "string",
            "format": "date",
            "title": "Fecha Verificacion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "institucion_id",
          "codigo",
          "nombre",
          "provincia",
          "tipo",
          "latitud",
          "longitud",
          "fuente",
          "fecha_verificacion"
        ],
        "title": "CompetitorOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InstitutionCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nombre"
          },
          "es_propia": {
            "type": "boolean",
            "title": "Es Propia",
            "default": false
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "title": "Color",
            "default": "#808080"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nombre"
        ],
        "title": "InstitutionCreate"
      },
      "InstitutionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nombre"
          },
          "es_propia": {
            "type": "boolean",
            "title": "Es Propia",
            "default": false
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "title": "Color",
            "default": "#808080"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "nombre"
        ],
        "title": "InstitutionOut"
      },
      "LocationType": {
        "type": "string",
        "enum": [
          "ATM",
          "SUCURSAL"
        ],
        "title": "LocationType"
      },
      "MarketCreate": {
        "properties": {
          "institucion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Institucion Id"
          },
          "provincia": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Provincia"
          },
          "fecha": {
            "type": "string",
            "format": "date",
            "title": "Fecha"
          },
          "total_atms": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Total Atms"
          },
          "total_sucursales": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Total Sucursales"
          },
          "depositos": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Depositos"
          },
          "prestamos": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Prestamos"
          },
          "transacciones": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Transacciones"
          },
          "utilizacion_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Utilizacion Pct"
          },
          "fuente": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Fuente"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "institucion_id",
          "provincia",
          "fecha",
          "total_atms",
          "total_sucursales",
          "depositos",
          "prestamos",
          "transacciones",
          "fuente"
        ],
        "title": "MarketCreate"
      },
      "MarketOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "institucion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Institucion Id"
          },
          "provincia": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Provincia"
          },
          "fecha": {
            "type": "string",
            "format": "date",
            "title": "Fecha"
          },
          "total_atms": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Total Atms"
          },
          "total_sucursales": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Total Sucursales"
          },
          "depositos": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
            "title": "Depositos"
          },
          "prestamos": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,16}|(?=[\\d.]{1,19}0*$)\\d{0,16}\\.\\d{0,2}0*$)",
            "title": "Prestamos"
          },
          "transacciones": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Transacciones"
          },
          "utilizacion_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Utilizacion Pct"
          },
          "fuente": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Fuente"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "institucion_id",
          "provincia",
          "fecha",
          "total_atms",
          "total_sucursales",
          "depositos",
          "prestamos",
          "transacciones",
          "fuente"
        ],
        "title": "MarketOut"
      },
...
      "Page_CompetitorOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompetitorOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[CompetitorOut]"
      },
...
      "Page_InstitutionOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstitutionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[InstitutionOut]"
      },
      "Page_MarketOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MarketOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MarketOut]"
      },
...
      "Series_Comparison_": {
        "properties": {
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Comparison"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "items"
        ],
        "title": "Series[Comparison]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}